refactor!: create `data.tabular` and `data.tabular.typing` modules
