+++ Runtime/safe-ds/safeds/data/__init__.py
@@ -1,17 +1,5 @@
-from ._column import Column, ColumnStatistics
-from ._column_type import (
-    BooleanColumnType,
-    ColumnType,
-    FloatColumnType,
-    IntColumnType,
-    OptionalColumnType,
-    StringColumnType,
-)
 from ._imputer import Imputer
 from ._label_encoder import LabelEncoder
 from ._one_hot_encoder import OneHotEncoder
 from ._ordinal_encoder import OrdinalEncoder
-from ._row import Row
 from ._supervised_dataset import SupervisedDataset
-from ._table import Table
-from ._table_schema import TableSchema
