+++ Runtime/safe-ds/safeds/data/_ordinal_encoder.py
@@ -1,5 +1,5 @@
 from safeds import exceptions
-from safeds.data._table import Table
+from safeds.data.tabular import Table
 from sklearn import preprocessing
