+++ Runtime/safe-ds/safeds/data/tabular/typing/__init__.py
@@ -1,0 +1,9 @@
+from ._column_type import (
+    BooleanColumnType,
+    ColumnType,
+    FloatColumnType,
+    IntColumnType,
+    OptionalColumnType,
+    StringColumnType,
+)
+from ._table_schema import TableSchema
