+++ Runtime/safe-ds/safeds/data/tabular/typing/_column_type.py
