+++ Runtime/safe-ds/safeds/data/tabular/typing/_table_schema.py
@@ -5,7 +5,7 @@ from dataclasses import dataclass
 import pandas as pd
 from safeds.exceptions import UnknownColumnNameError
 
-from ._column_type import ColumnType
+from data.tabular.typing._column_type import ColumnType
 
 
 @dataclass
